# Optimized linear combination of points