The goal of this PR is to optimize a commonly used operation `(x * k + y * l)` where `x` and `y` are curve points, and `k` and `l` are scalars. It is currently used in `verify.rs` and `recoverable.rs`. 

On my machine it speeds up ECDSA verification from 176us to 145us (and the newly added benchmark shows speed-up from 157us to 124us for linear combination alone).

Currently there PR contains a `lincomb_generic()` that takes arrays, and it has two aliases: `mul()` (internal, used in operator traits) and `lincomb()` (public, for two-point linear combinations).