@hdevalence 2021-07-14
Two traits that could be useful as design inspiration are the `curve25519-dalek` [`MultiscalarMul`](https://docs.rs/curve25519-dalek-ng/4.0.1/curve25519_dalek_ng/traits/trait.MultiscalarMul.html) and [`VartimeMultiscalarMul`](https://docs.rs/curve25519-dalek-ng/4.0.1/curve25519_dalek_ng/traits/trait.VartimeMultiscalarMul.html).

IME there's pretty minimal cost to making an arbitrary-length multiscalar multiplication API over a double-base scalar multiplication API.  Basically the only downside is that the lookup tables have to be heap-allocated, but this is a pretty minimal cost even for the double-base case.

This is just my opinion, but I think it's preferable to have *only* the arbitrary-length version, for reduced API surface.  If you're going to have a specialized double-base scmul function, it's useful to have it hardcode one of the points (so it can use a bigger, statically precomputed lookup table).