@tarcieri 2021-07-15
@fjarri hope you don't mind but I flipped this PR out of draft

I think it's a reasonable self-contained first step.