@fjarri 2021-07-16
I just wanted to emphasize that it shouldn't be merged yet :)

So, I added a generic version of `lincomb` that could potentially replace `mul_windowed` and `lincomb`. The performance is the following:

high-level operations/lincomb via mul+add [157.31 us 157.39 us 157.48 us]
high-level operations/lincomb optimized [124.71 us 125.17 us 125.64 us]
high-level operations/lincomb optimized (generic) [122.25 us 122.31 us 122.37 us]
high-level operations/mul_windowed [80.548 us 80.767 us 80.993 us]
high-level operations/mul_windowed (generic) [80.870 us 81.071 us 81.269 us]

A strange effect is that `ecdsa/verify_prehashed` benchmark is ~5us faster with a non-generic version (152 vs 147us), while the benchmark of `lincomb` itself shows that the generic version is slightly faster.

Do you think the generic version overcomplicates things?