@tarcieri 2021-07-16
I wouldn't worry too much about a modicum of performance lost to the generic version.

> I wish Rust had maps and zips defined for static arrays...

It *should* retain all of the information necessary to optimize those as part of `slice::Iter`, which is an `ExactSizeIterator` over a fix-sized structure. Though I have seen LLVM optimize those poorly in the past and we've often used for loops over a static range instead.