@fjarri 2021-07-16
> It should retain all of the information necessary to optimize those as part of slice::Iter, which is an ExactSizeIterator over a fix-sized structure.

Yes, but as far as I know it is not possible to `collect()` into an array. I guess I can try to rewrite it in lazy maps, but this will result in some freaky expressions.