@tarcieri 2021-07-16
> Yes, but as far as I know it is not possible to collect() into an array. 

Aah yes, that's been the long-desired one which continues to be an ongoing source of controversy as to how to implement it. I think this is the most up-to-date tracking issue on where things are at with that:

https://github.com/rust-lang/rust/issues/81615