@fjarri 2021-07-16
So, do you think it's ok to use the current approach with `static_map`, noting that they should be replaced with standard tools when those are available? I made an attempt at rewriting the function in terms of maps and folds, but it's pretty hard to do without re-calculating intermediate values. And even if I manage to do it, the result will be quite obscure - I'd rather keep the imperative logic flow.