@fjarri 2021-07-16
Ok, done. I just exposed it publicly as `lincomb()` (because benchmarks need it). Ideally, I think, the feature `expose-field` should be renamed to something like `expose-internals` and use to gate stuff like that to expose it to benchmakrs. But I'm not sure if I should do that in this PR.

As for a trait that @hdevalence suggested, I think it's a good idea, but I don't know how I feel about using iterators instead of fixed-size arrays. Maybe we can leave this PR at just a function and add a trait next (sometime before the next release).