@tarcieri 2021-07-18
> Ideally, I think, the feature expose-field should be renamed to something like expose-internals and used to gate stuff like that to expose it to benchmakrs. But I'm not sure if I should do that in this PR.

There are various and somewhat inconsistent ways that low-level internals are exposed across crates in the RustCrypto org.

- Some crates have a `hazmat` feature (e.g. `aes`)
- `chacha20` and `salsa20` have `expose-core`

It might be good to come up with a single naming convention for this, but doing so is a breaking change. I can try to open a sort of meta issue on this so we can pick a single name for this sort of thing and use it consistently.