@tarcieri 2021-07-18
@fjarri thanks for implementing this!

Would definitely be interested in precomputed powers for the generator point if you're interested in implementing that as well:

> in verify.rs and recoverable.rs one of the points of the linear combination is a generator. So it can be sped up even more by pre-calculating generator powers.