@tarcieri 2021-07-18
> the biggest problem here is precomputing and storing the required table.

Any thoughts on `const fn` (and possibly const generics) maturity for computing them at compile time with miri?

We could feature gate it so it doesn't impact embedded targets.

> ...we can use regular table lookup instead of constant-time lookup.

That'd be great for ECDSA. Also generally I've wondered if wNAF would be nice here, although I'm not sure if the benefits are a wash given secp256k1 endomorphisms. Are you familiar with what bitcoin-core/libsecp256k1 does here? (I am not, perhaps I should look)